Implement summoning pity rates
- Refactor summoning odds calculation into static pure functions to make
  them easier to test
- Add new pity counter column to DbPlayerBannerData

This does not activate the pity rate calculation for actual results;
everyone will have a counter of 0 which will not accumulate.

The algorithm for pity rates is somewhat empirical based on a random
screenshot and may not distribute the .5% increases in the exact way
that the game originally did.
